pyproject.toml
[tool.poetry]
name = "target-kafka"
version = "0.1.1"
description = "`target-kafka` is a Singer target for Kafka, built with the Meltano Singer SDK."
readme = "README.md"
authors = ["Reid LAI <reidlai@gmail.com>"]
keywords = [
    "ELT",
    "Kafka",
]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
license = "Apache-2.0"
packages = [
    { include = "target_kafka" },
]

[tool.poetry.dependencies]
python = ">=3.8"
singer-sdk = { version="~=0.37.0", extras = ["faker"] }
fs-s3fs = { version = "~=1.1.1", optional = true }
requests = "~=2.31.0"
kafka-python = "^2.0.2"

[tool.poetry.dev-dependencies]
pytest = ">=7.4.0"
singer-sdk = { version="~=0.37.0", extras = ["testing"] }

[tool.poetry.extras]
s3 = ["fs-s3fs"]

[tool.ruff]
src = ["target_kafka"]
target-version = "py38"

[tool.ruff.lint]
ignore = [
    "ANN101",  # missing-type-self
    "ANN102",  # missing-type-cls
    "COM812",  # missing-trailing-comma
    "ISC001",  # single-line-implicit-string-concatenation
]
select = ["ALL"]

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.lint.isort]
known-first-party = ["target_kafka"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[build-system]
requires = ["poetry-core==1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
# CLI declaration
target-kafka = 'target_kafka.target:Targetkafka.cli'
